move classes into separated modules
