+++ src/cotletico_manager.py
@@ -1,74 +1,11 @@
 import random
 import json
-
-
-class Player:
-    def __init__(self, name, surname, position, rating, birthdate, number, team, statistics):
-        self.name = name
-        self.surname = surname
-        self.position = position
-        self.rating = rating
-        self.birthdate = birthdate
-        self.number = number
-        self.team = team
-        self.statistics = statistics
-
-    def update_statistics(self, statistics):
-        self.statistics = statistics
-
-    def __lt__(self, other):
-        return self.rating < other.rating
-
-
-class Team:
-    def __init__(self, name, players, matches=[], flag=None, coach=''):
-        self.name = name
-        self.flag = flag
-        self.players = players
-        self.matches = matches
-        self.coach = coach
-
-    def add_match(self, match):
-        self.matches.append(match)
-
-
-class Match:
-    def __init__(self, stadium, attendance, date, host, guest, events, statistics):
-        self.stadium = stadium
-        self.attendance = attendance
-        self.date = date
-        self.host = host
-        self.guest = guest
-        self.events = events
-        self.statistics = statistics
-
-    def add_event(self, event):
-        self.events.append(event)
-
-    def update_statistics(self, statistics):
-        self.statistics = statistics
-
-
-class Stadium:
-    def __init__(self, name, city, capacity, image):
-        self.name = name
-        self.city = city
-        self.players = capacity
-        self.image = image
-
-
-class Event:
-    def __init__(self, type, time):
-        self.type = type
-        self.time = time
-
-
-class Schedule:
-    def __init__(self, matches):
-        self.matches = matches
-
-    def add_match(self, match):
-        self.matches.append(match)
+import player
+import team
+import match
+import stadium
+import event
+import schedule
 
 
 def teams_from_json(filename: str):
@@ -81,11 +18,11 @@ def teams_from_json(filename: str):
     for team_name in data.keys():
         players = list()
         for player_info in data[team_name]:
-            player = Player(**player_info)
-            players.append(player)
-            player.statistics = dict({'injured': False, 'red': False})
-        team = Team(team_name, players)
-        teams.append(team)
+            pl = player.Player(**player_info)
+            players.append(pl)
+            pl.statistics = dict({'injured': False, 'red': False})
+        tm = team.Team(team_name, players)
+        teams.append(tm)
     return teams
