+++ src/event.py
@@ -1,0 +1,4 @@
+class Event:
+    def __init__(self, type, time):
+        self.type = type
+        self.time = time
