+++ src/schedule.py
@@ -1,0 +1,6 @@
+class Schedule:
+    def __init__(self, matches):
+        self.matches = matches
+
+    def add_match(self, match):
+        self.matches.append(match)
