+++ src/stadium.py
@@ -1,0 +1,6 @@
+class Stadium:
+    def __init__(self, name, city, capacity, image):
+        self.name = name
+        self.city = city
+        self.players = capacity
+        self.image = image
